first use of int64_t. int64_t guarantees we get 64 bits or 8 bytes.

+++ main.cpp
@@ -3,7 +3,7 @@
 // C++ Primitive Data Types | Modern Cpp Series Ep. 10
 
 int main() {
-    long long x = 4222222212121221212;
+    int64_t x = 4222222212121221212;
     std::cout << x << std::endl;
     std::cout << sizeof(x) << std::endl;
     return 0;
